+++ example/react-native.config.js
@@ -1,0 +1,9 @@
+const path = require('path');
+
+module.exports = {
+  dependencies: {
+    'react-native-telematics': {
+      root: path.resolve(__dirname, '..'),
+    },
+  },
+};